requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
Django==3.0.8
django-jet2==1.0.12
django-random-queryset==0.1.3
djangorestframework==3.11.0
idna==2.10
pycparser==2.20
pyTelegramBotAPI==3.7.2
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
